Fix Eruption GUI
Fix multiple minor bugs in Eruption GUI

+++ eruption-fx-proxy/src/util.rs
@@ -56,7 +56,7 @@ pub fn process_image_buffer(
     for x in 0..num_cols {
         for y in 0..num_rows {
             let key_index: usize = *rows_topology
-                .get(x + (y * (num_cols + 1)) as usize)
+                .get(x + (y * (num_cols + 1)))
                 .unwrap_or(&0) as usize
                 + 1;
